move pristine to separate package
