+++ internal/commands/show.go
@@ -26,7 +26,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/splunk/qbec/internal/model"
 	"github.com/splunk/qbec/internal/objsort"
-	"github.com/splunk/qbec/internal/remote"
+	"github.com/splunk/qbec/internal/pristine"
 	"github.com/splunk/qbec/internal/sio"
 	"github.com/splunk/qbec/internal/types"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -182,7 +182,7 @@ func doShow(args []string, config showCommandConfig) error {
 	}
 
 	for _, o := range objects {
-		p := remote.QbecPristine{}
+		p := pristine.QbecPristine{}
 		ret, _ := p.CreateFromPristine(o)
 		displayObjects = append(displayObjects, mapper(ret))
 	}
